[15] Convert view structure to use Blueprint
